+++ src/extension.ts
@@ -1,6 +1,6 @@
 'use strict';
 import { ExtensionContext, commands, window, workspace, TextEditor, TextDocument } from 'vscode';
-// import * as vscode from 'vscode';
+import * as vscode from 'vscode';
 import { IConfig, IRange } from './types';
 
 export function activate(ctx: ExtensionContext) {
@@ -13,31 +13,37 @@ export function activate(ctx: ExtensionContext) {
 			config.allowedNumberOfEmptyLines = 0;
 		}
 	}
-	function removeEmptyLines(inSelection: boolean): void {
+
+	function removeEmptyLines(inSelection: boolean, editor: vscode.TextEditor): void {
 		updateConfig();
 
-		const { activeTextEditor } = window;
-		if (!activeTextEditor) {
-			return;
-		}
-		const { document } = activeTextEditor;
+		const { document } = editor;
 
 		if (inSelection) {
-			const { selections } = activeTextEditor;
+			const { selections } = editor;
+
 			if (selections.length === 1 && selections[0].isEmpty) {
+				// remove all adjacent empty lines up & down of the cursor
 				const tempAllowedNumberOfEmptyLines = config.allowedNumberOfEmptyLines;
 				config.allowedNumberOfEmptyLines = 0;
-				removeEmptyLinesInRange(activeTextEditor, document, [[selections[0].start.line, selections[0].start.line + 1]]);
+				const activeLine = document.lineAt(selections[0].start.line);
+				if (!activeLine.isEmptyOrWhitespace) {
+					return;
+				} else {
+					const closestUp = findUpClosestNonEmptyLine(selections[0].start.line, document);
+					const closestDown = findDownClosestNonEmptyLine(selections[0].start.line, document);
+					removeEmptyLinesInRange(editor, document, [[closestUp, closestDown]]);
+				}
 				config.allowedNumberOfEmptyLines = tempAllowedNumberOfEmptyLines;
 			} else {
 				const ranges: IRange[] = [];
 				selections.forEach((selection) => {
 					ranges.push([selection.start.line, selection.end.line]);
 				});
-				removeEmptyLinesInRange(activeTextEditor, document, ranges);
+				removeEmptyLinesInRange(editor, document, ranges);
 			}
 		} else {
-			removeEmptyLinesInRange(activeTextEditor, document, [[0, document.lineCount - 1]]);
+			removeEmptyLinesInRange(editor, document, [[0, document.lineCount - 1]]);
 		}
 	}
 	function removeEmptyLinesInRange(editor: TextEditor, document: TextDocument, ranges: IRange[]): void {
@@ -58,13 +64,25 @@ export function activate(ctx: ExtensionContext) {
 			});
 		});
 	}
+	function findUpClosestNonEmptyLine(line: number, document: vscode.TextDocument) {
+		for (let i = line; i > 0; i--) {
+			const line = document.lineAt(i);
+			if (line.isEmptyOrWhitespace) { continue; }
+			return i;
+		}
+		return 0;
+	}
+	function findDownClosestNonEmptyLine(line: number, document: vscode.TextDocument) {
+		for (let i = line; i < document.lineCount; i++) {
+			const line = document.lineAt(i);
+			if (line.isEmptyOrWhitespace) { continue; }
+			return i;
+		}
+		return document.lineCount - 1;
+	}
 
-	const removeEmptyLinesInDocument = commands.registerCommand(`${extName}.inDocument`, () => {
-		removeEmptyLines(false);
-	});
-	const removeEmptyLinesInSelection = commands.registerCommand(`${extName}.inSelection`, () => {
-		removeEmptyLines(true);
-	});
+	const removeEmptyLinesInDocument = commands.registerTextEditorCommand(`${extName}.inDocument`, removeEmptyLines.bind(null, false));
+	const removeEmptyLinesInSelection = commands.registerTextEditorCommand(`${extName}.inSelection`, removeEmptyLines.bind(null, true));
 	ctx.subscriptions.push(removeEmptyLinesInDocument, removeEmptyLinesInSelection);
 }