+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3 `06 Feb 2019`
+
+- ✨ Remove all adjacent empty lines when executing `remove-empty-lines.inSelection` without selection
+- 🔨 Reduce the size of `.gif` and improve their quality **705Kb** => **121Kb**
+
 ## 0.0.2 `29 Dec 2018`
 
 - ✨ Added setting `remove-empty-lines.allowedNumberOfEmptyLines`